+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 Unreleased
 ----------
 
+- [code gen] Add check that presenter function return types return a `UiState` type.
 - [code gen] Drop KSP1 support.
 
 0.31.0